requirements.txt
flask>=3.0.0
requests>=2.31.0
google-cloud-texttospeech>=2.27.0
pydub>=0.25.1
tqdm>=4.65.0
moviepy>=1.0.3
pillow>=10.0.0
werkzeug>=3.0.0 